+++ ui/admin/test/server.ts
@@ -1,6 +1,6 @@
 import { RequestHandler } from "msw";
 import { setupServer } from "msw/node";
-import { defaultMockedHandlers } from "test/__mocks__/handlers";
+import { defaultMockedHandlers } from "test/mocks/handlers/default";
 
 // Setup requests interception using the given handlers
 export const server = setupServer(...defaultMockedHandlers);
